Added mechanism to designate arbitrary exceptions as recoverable.
This important feature is needed because, although we have a mechanism
for marking exceptions that extend RecoverableConnectorException as
recoverable at the connector level, exception may need to be treated as
recoverable at the resource level too. Resource implementations may not
always have the luxury of creating their own exception types when
relying on third party libraries.