Changed (Async)PorterRecords semantics to always run generators to first suspension point.
This was always a side-effect of the (previous, coroutine) async
behaviour, but since fibers, is no longer a necessary side-effect. Yet,
some libraries may depend on this behaviour, such as those expecting to
resolve a deferred before the first suspension point. We now adopt these
semantics intentionally and replicate them to the sync flow.